+++ packages/examples/src/clangd/client/config.ts
@@ -57,6 +57,7 @@ export const createWrapperConfig = async (config: {
         },
         vscodeApiConfig: {
             enableTextmate: true,
+            loadThemes: true,
             serviceOverrides: {
                 ...getConfigurationServiceOverride(),
                 ...getKeybindingsServiceOverride(),