+++ packages/examples/src/eclipse.jdt.ls/client/main.ts
@@ -25,6 +25,7 @@ export const runEclipseJdtLsClient = () => {
         htmlContainer: document.getElementById('monaco-editor-root')!,
         vscodeApiConfig: {
             enableTextmate: true,
+            loadThemes: true,
             serviceOverrides: {
                 ...getKeybindingsServiceOverride(),
             },