+++ packages/examples/src/groovy/client/main.ts
@@ -21,6 +21,7 @@ const userConfig: WrapperConfig = {
     htmlContainer: document.getElementById('monaco-editor-root')!,
     vscodeApiConfig: {
         enableTextmate: true,
+        loadThemes: true,
         serviceOverrides: {
             ...getKeybindingsServiceOverride(),
         },