+++ packages/examples/src/langium/langium-dsl/config/extendedConfig.ts
@@ -31,6 +31,7 @@ export const setupLangiumClientExtended = async (): Promise<WrapperConfig> => {
         htmlContainer: document.getElementById('monaco-editor-root')!,
         vscodeApiConfig: {
             enableTextmate: true,
+            loadThemes: true,
             serviceOverrides: {
                 ...getKeybindingsServiceOverride()
             },