+++ packages/examples/src/multi/twoLanguageClients.ts
@@ -42,6 +42,7 @@ print("Hello Moon!")
         htmlContainer: document.getElementById('monaco-editor-root')!,
         vscodeApiConfig: {
             enableTextmate: true,
+            loadThemes: true,
             serviceOverrides: {
                 ...getKeybindingsServiceOverride()
             },