+++ packages/examples/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this npm module are documented in this file.
 ## [2024.10.6] - 2024-11-13
 
 - Updated to `monaco-languageclient@9.0.0-next.7`, `monaco-editor-wrapper@6.0.0-next.7` and `@typefox/monaco-editor-react@6.0.0-next.7`.
-- Updated all `@codingame/monaco-vscode` packages to `11.0.1`.
+- Updated all `@codingame/monaco-vscode` packages to `11.0.2`.
 
 ## [2024.10.5] - 2024-10-31